Remove translator dependency in libs/common #1767
