+++ libs/translator/reader/utils.h
@@ -44,51 +44,6 @@ class UsdArnoldReaderContext;
 
 #include "timesettings.h"
 
-// TODO: primvarsremapper and inputattribute classes should be moved out from this file
-class PrimvarsRemapper
-{
-public:
-    PrimvarsRemapper() {}
-    virtual ~PrimvarsRemapper() {}  
-
-    virtual bool RemapValues(const UsdGeomPrimvar &primvar, const TfToken &interpolation, 
-        VtValue &value);
-    virtual bool RemapIndexes(const UsdGeomPrimvar &primvar, const TfToken &interpolation, 
-        std::vector<unsigned int> &indexes);
-    virtual void RemapPrimvar(TfToken &name, std::string &interpolation);
-};
-
-struct InputAttribute {
-    InputAttribute(const UsdAttribute& attribute) : attr(attribute), primvar(nullptr) {}
-    InputAttribute(const UsdGeomPrimvar& primv) : attr(primv.GetAttr()), primvar(&primv) {}
-
-    const UsdAttribute& GetAttr() { return attr; }
-
-    bool Get(VtValue* value, double frame) const
-    {
-        bool res = false;
-        if (primvar) {
-            if (computeFlattened)
-                res = primvar->ComputeFlattened(value, frame);
-            else
-                res = primvar->Get(value, frame);
-        } else
-            res = attr.Get(value, frame);
-
-        if (primvar && primvarsRemapper)
-            primvarsRemapper->RemapValues(*primvar, primvarInterpolation, *value);
-        return res;
-
-    }
-
-    const UsdAttribute& attr;
-    const UsdGeomPrimvar* primvar;
-    bool computeFlattened = false;
-    PrimvarsRemapper *primvarsRemapper = nullptr;
-    TfToken primvarInterpolation;
-
-};
-
 /** Read Xformable transform as an arnold shape "matrix"
  */
 void ReadMatrix(const UsdPrim& prim, AtNode* node, const TimeSettings& time, 
@@ -99,295 +54,7 @@ AtArray *ReadMatrix(const UsdPrim& prim, const TimeSettings& time,
 
 AtArray *ReadLocalMatrix(const UsdPrim &prim, const TimeSettings &time);
 
-/** Read String arrays, and handle the conversion from std::string / TfToken to AtString.
- */
-inline
-size_t ReadStringArray(UsdAttribute attr, AtNode *node, const char *attrName, const TimeSettings &time)
-{
-    // Strings can be represented in USD as std::string, TfToken or SdfAssetPath.
-    // We'll try to get the input attribute value as each of these types
-    VtArray<std::string> arrayStr;
-    VtArray<TfToken> arrayToken;
-    VtArray<SdfAssetPath> arrayPath;
-    AtArray *outArray = nullptr;
-    size_t size;
 
-    if (attr.Get(&arrayStr, time.frame)) {
-        size = arrayStr.size();
-        if (size > 0) {
-            outArray = AiArrayAllocate(size, 1, AI_TYPE_STRING);
-            for (size_t i = 0; i < size; ++i) {
-                if (!arrayStr[i].empty())
-                    AiArraySetStr(outArray, i, AtString(arrayStr[i].c_str()));
-                else
-                    AiArraySetStr(outArray, i, AtString(""));
-            }
-        }
-    } else if (attr.Get(&arrayToken, time.frame)) {
-        size = arrayToken.size();
-        if (size > 0) {
-            outArray = AiArrayAllocate(size, 1, AI_TYPE_STRING);
-            for (size_t i = 0; i < size; ++i) {
-                if (!arrayToken[i].GetString().empty())
-                    AiArraySetStr(outArray, i, AtString(arrayToken[i].GetText()));
-                else
-                    AiArraySetStr(outArray, i, AtString(""));
-            }
-        }
-    } else if (attr.Get(&arrayPath, time.frame)) {
-        size = arrayPath.size();
-        if (size > 0) {
-            outArray = AiArrayAllocate(size, 1, AI_TYPE_STRING);
-            for (size_t i = 0; i < size; ++i) {
-                if (!arrayPath[i].GetResolvedPath().empty())
-                    AiArraySetStr(outArray, i, AtString(arrayPath[i].GetResolvedPath().c_str()));
-                else
-                    AiArraySetStr(outArray, i, AtString(""));
-            }
-        }
-    }
-
-    if (outArray)
-        AiNodeSetArray(node, AtString(attrName), outArray);
-    else
-        AiNodeResetParameter(node, AtString(attrName));
-
-    return 1; // return the amount of motion keys
-}
-
-template <class U, class A>
-size_t ReadArray(
-    UsdAttribute attr, AtNode* node, const char* attrName, const TimeSettings& time, uint8_t attrType = AI_TYPE_NONE)
-{
-    InputAttribute inputAttr(attr);
-    return ReadArray<U, A>(inputAttr, node, attrName, time, attrType);
-}
-/** Convert a USD array attribute (type U), to an Arnold array (type A).
- *  When both types are identical, we can simply their pointer to create the
- *array. Otherwise we need to copy the data first
- **/
-template <class U, class A>
-size_t ReadArray(
-    InputAttribute& attr, AtNode* node, const char* attrName, const TimeSettings& time, uint8_t attrType = AI_TYPE_NONE)
-{
-    bool sameData = std::is_same<U, A>::value;
-    const UsdAttribute& usdAttr = attr.GetAttr();
-
-    if (attrType == AI_TYPE_NONE) {
-        if (std::is_same<A, float>::value)
-            attrType = AI_TYPE_FLOAT;
-        else if (std::is_same<A, int>::value)
-            attrType = AI_TYPE_INT;
-        else if (std::is_same<A, bool>::value)
-            attrType = AI_TYPE_BOOLEAN;
-        else if (std::is_same<A, unsigned int>::value)
-            attrType = AI_TYPE_UINT;
-        else if (std::is_same<A, unsigned char>::value)
-            attrType = AI_TYPE_BYTE;
-        else if (std::is_same<A, GfVec3f>::value)
-            attrType = AI_TYPE_VECTOR;
-        else if (std::is_same<A, AtRGB>::value)
-            attrType = AI_TYPE_RGB;
-        else if (std::is_same<A, AtRGBA>::value)
-            attrType = AI_TYPE_RGBA;
-        else if (std::is_same<A, GfVec4f>::value)
-            attrType = AI_TYPE_RGBA;
-        else if (std::is_same<A, TfToken>::value)
-            attrType = AI_TYPE_STRING;
-        else if (std::is_same<A, std::string>::value)
-            attrType = AI_TYPE_STRING;
-        else if (std::is_same<A, GfMatrix4f>::value)
-            attrType = AI_TYPE_MATRIX;
-        else if (std::is_same<A, GfMatrix4d>::value)
-            attrType = AI_TYPE_MATRIX;
-        else if (std::is_same<A, AtMatrix>::value) {
-            if (std::is_same<U, GfMatrix4f>::value)
-                sameData = true;
-            attrType = AI_TYPE_MATRIX;
-        } else if (std::is_same<A, AtVector>::value) {
-            attrType = AI_TYPE_VECTOR;
-            if (std::is_same<U, GfVec3f>::value) // AtVector is represented the same
-                                                 // way as GfVec3f
-                sameData = true;
-        } else if (std::is_same<A, GfVec2f>::value)
-            attrType = AI_TYPE_VECTOR2;
-        else if (std::is_same<A, AtVector2>::value) {
-            attrType = AI_TYPE_VECTOR2;
-            if (std::is_same<U, GfVec2f>::value) // AtVector2 is represented the
-                                                 // same way as GfVec2f
-                sameData = true;
-        }
-    }
-
-    // Call a dedicated function for string conversions
-    if (attrType == AI_TYPE_STRING)
-        return ReadStringArray(usdAttr, node, attrName, time);
-
-    bool animated = time.motionBlur && usdAttr.ValueMightBeTimeVarying();
-
-    if (!animated) {
-        // Single-key arrays
-        VtValue val;
-        if (!attr.Get(&val, time.frame)) {
-            // Create an empty array
-            AiNodeSetArray(node, AtString(attrName), AiArrayConvert(0, 1, attrType, nullptr));
-            return 0;
-        }
-
-
-        const VtArray<U>* array = &(val.Get<VtArray<U>>());
-
-        size_t size = array->size();
-        if (size > 0) {
-            if (std::is_same<U, GfMatrix4d>::value) {
-                // special case for matrices. They're single
-                // precision in arnold but double precision in USD,
-                // and there is no copy from one to the other.
-                VtArray<GfMatrix4d>* arrayMtx = (VtArray<GfMatrix4d>*)(array);
-                GfMatrix4d* matrices = arrayMtx->data();
-                std::vector<AtMatrix> arnoldVec(size);
-                for (size_t v = 0; v < size; ++v) {
-                    AtMatrix& aiMat = arnoldVec[v];
-                    const double* matArray = matrices[v].GetArray();
-                    for (unsigned int i = 0; i < 4; ++i)
-                        for (unsigned int j = 0; j < 4; ++j)
-                            aiMat[i][j] = matArray[4 * i + j];
-                }
-                AiNodeSetArray(node, AtString(attrName), AiArrayConvert(size, 1, AI_TYPE_MATRIX, &arnoldVec[0]));
-
-            } else if (sameData) {
-                // The USD data representation is the same as the Arnold one, we don't
-                // need to convert the data
-                AiNodeSetArray(node, AtString(attrName), AiArrayConvert(size, 1, attrType, array->cdata()));
-            } else {
-                // Different data representation between USD and Arnold, we need to
-                // copy the vector. 
-                VtArray<A> arnold_vec;
-                arnold_vec.reserve(array->size());
-                for (const auto &elem : (*array))
-                    arnold_vec.push_back(static_cast<A>(elem));
-                
-                AiNodeSetArray(node, AtString(attrName), AiArrayConvert(size, 1, attrType, arnold_vec.cdata()));
-            }
-        } else
-            AiNodeResetParameter(node, AtString(attrName));
-
-        return 1; // return the amount of keys
-    } else {
-        // Animated array
-        GfInterval interval(time.start(), time.end(), false, false);
-        std::vector<double> timeSamples;
-        usdAttr.GetTimeSamplesInInterval(interval, &timeSamples);
-        // need to add the start end end keys (interval has open bounds)
-        size_t numKeys = timeSamples.size() + 2;
-
-        double timeStep = double(interval.GetMax() - interval.GetMin()) / double(numKeys - 1);
-        double timeVal = interval.GetMin();
-
-        VtValue val;
-        if (!attr.Get(&val, timeVal)) {
-            // Create an empty array
-            AiNodeSetArray(node, AtString(attrName), AiArrayConvert(0, 1, attrType, nullptr));
-            return 0;
-        }
-
-        const VtArray<U>* array = &(val.Get<VtArray<U>>());
-
-        // Arnold arrays don't support varying element counts per key.
-        // So if we find that the size changes over time, we will just take a single key for the current frame        
-        size_t size = array->size();
-        if (size == 0)
-            return 0;
-        
-        if (std::is_same<U, GfMatrix4d>::value) {
-            VtArray<AtMatrix> arnoldVec(size * numKeys);
-            int index = 0;
-
-            for (size_t i = 0; i < numKeys; i++, timeVal += timeStep) {
-                if (i > 0) {
-                    // if a time sample is missing, we can't translate 
-                    // this attribute properly
-                    if (!attr.Get(&val, timeVal))
-                        return 0;
-                    
-                    array = &(val.Get<VtArray<U>>());
-                }
-                VtArray<GfMatrix4d>* arrayMtx = (VtArray<GfMatrix4d>*)(array);
-                GfMatrix4d* matrices = arrayMtx->data();
-                if (arrayMtx->size() != size) {
-                    // Arnold won't support varying element count. 
-                    // We need to only consider a single key corresponding to the current frame
-                    arnoldVec.clear();
-                    if (!attr.Get(&val, time.frame)) 
-                        return 0;
-                
-                    index = 0;
-                    array = &(val.Get<VtArray<U>>());
-                    size = array->size(); // update size to the current frame one
-                    if (size == 0)
-                        return 0;
-                
-                    numKeys = 1; // we just want a single key
-                    arnoldVec.resize(size);
-                    arrayMtx = (VtArray<GfMatrix4d>*)(array);
-                    matrices = arrayMtx->data();
-                    i = numKeys; // this will stop the "for" loop
-                }
-
-                for (size_t v = 0; v < size; ++v, ++index) {
-                    AtMatrix& aiMat = arnoldVec[index];
-                    const double* matArray = matrices[v].GetArray();
-                    for (unsigned int k = 0; k < 4; ++k)
-                        for (unsigned int j = 0; j < 4; ++j)
-                            aiMat[k][j] = matArray[4 * k + j];
-                }
-            }
-            if (size > 0)
-                AiNodeSetArray(node, AtString(attrName), AiArrayConvert(size, numKeys, AI_TYPE_MATRIX, arnoldVec.data()));
-        } else {
-            A* arnoldVec = new A[size * numKeys], *ptr = arnoldVec;
-            for (size_t i = 0; i < numKeys; i++, timeVal += timeStep) {
-                if (i > 0) {
-                    // if a time sample is missing, we can't translate 
-                    // this attribute properly
-                    if (!attr.Get(&val, timeVal)) {
-                        size = 0;
-                        break;
-                    }
-                    array = &(val.Get<VtArray<U>>());
-                }
-                if (array->size() != size) {
-                     // Arnold won't support varying element count. 
-                    // We need to only consider a single key corresponding to the current frame
-                    if (!attr.Get(&val, time.frame)) {
-                        size = 0;
-                        break;
-                    }                        
-
-                    delete [] arnoldVec;
-                    array = &(val.Get<VtArray<U>>()); 
-                    size = array->size(); // update size to the current frame one
-                    numKeys = 1; // we just want a single key now
-                    // reallocate the array
-                    arnoldVec = new A[size * numKeys];
-                    ptr = arnoldVec;
-                    i = numKeys; // this will stop the "for" loop after the concatenation
-                    
-                }
-                for (unsigned j=0; j < array->size(); j++)
-                    *ptr++ = array->data()[j];
-            }
-
-            if (size > 0)
-                AiNodeSetArray(node, AtString(attrName), AiArrayConvert(size, numKeys, attrType, arnoldVec));
-            else
-                numKeys = 0;
-
-            delete [] arnoldVec;
-        }
-        return numKeys;
-    }
-}
 
 size_t ReadTopology(
     UsdAttribute& usdAttr, AtNode* node, const char* attrName, const TimeSettings& time, UsdArnoldReaderContext &context);
